requirements.txt
agate==1.9.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
black==25.1.0
bs4==0.0.2
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
daff==1.3.46
dbt-adapters==1.14.6
dbt-common==1.18
dbt-core==1.9.4
dbt-extractor==0.6.0
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
distlib==0.3.9
fastapi==0.115.12
filelock==3.18.0
flake8==7.2.0
greenlet==3.2.0
h11==0.14.0
identify==2.6.9
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
leather==0.4.0
logging==0.4.9.6
MarkupSafe==3.0.2
mashumaro==3.14
mccabe==0.7.0
more-itertools==10.6.0
msgpack==1.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.4
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
pre_commit==4.2.0
protobuf==5.29.4
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyee==12.1.1
pyflakes==3.3.2
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
text-unidecode==1.3
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.30.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.21.0
